feat:增加局域网联机功能

+++ GameCore.Soccer/Soccer/Messages.cs
@@ -10,7 +10,7 @@ namespace Soccer
     {
         public string serverName;
         public string serverAddress;
-        public int port;
+        public ushort port;
     }
     
     public enum IdentifierEnum
